Refactor: Separate component in SignUp page
