+++ src/pages/SignUp/components/SignUpForm/index.ts
@@ -1,0 +1,1 @@
+export * from './SignUpForm';
