+++ src/pages/SignUp/components/SignUpTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './SignUpTitle';
